Move useSWR utils to different folder
